[core] rm_rf(...): Ignore if doesn't exist.
(Closer behaviour to bash version.)

+++ lib_core/src/files/management.rs
@@ -51,6 +51,9 @@ pub fn rm_rf<P>(path: P) -> Result<(), CliError>
 where
     P: AsRef<Path>,
 {
+    if !path.as_ref().exists() {
+        return Ok(());
+    }
     std::fs::remove_dir_all(path).map_err(|e| IOError::with_debug(&e))?;
     Ok(())
 }
